Did a few fixes to improve creatures handling by computer player. Computer player now can use temple to heal disease, and puts much higher priority to training his creatures. Also, he will sacrifice creatures by mistake only in rare cases.

+++ keeperfx/src/globals.h
@@ -177,7 +177,7 @@ typedef short CrStateRet;
 /** Type which stores CrCkRet_* values. */
 typedef short CrCheckRet;
 /** Type which stores Job_* values. */
-typedef short CreatureJob;
+typedef unsigned long CreatureJob;
 /** Creature instance index, stores CrInst_* values. */
 typedef short CrInstance;
 /** Creature attack type, stores AttckT_* values. */
